chore(ReportDownload): Remove old PDF generator & feature flag
