+++ package.json
@@ -19,7 +19,6 @@
     "@redhat-cloud-services/frontend-components": "^6.1.0",
     "@redhat-cloud-services/frontend-components-config-utilities": "^4.4.1",
     "@redhat-cloud-services/frontend-components-notifications": "^6.1.0",
-    "@redhat-cloud-services/frontend-components-pdf-generator": "^4.0.7",
     "@redhat-cloud-services/frontend-components-remediations": "^4.0.11",
     "@redhat-cloud-services/frontend-components-utilities": "^6.0.0",
     "@redhat-cloud-services/javascript-clients-shared": "^2.0.0",
