+++ src/Routes.js
@@ -53,19 +53,6 @@ const reportsRoutes = [
     component: lazy(
       () =>
         import(
-          /* webpackChunkName: "ReportDetails" */ 'SmartComponents/ReportDownload/ReportDownload'
-        ),
-    ),
-  },
-  {
-    path: 'reports/:report_id/pdfv2',
-    title: `Export report - ${defaultReportTitle}`,
-    requiredPermissions: [...defaultPermissions, 'compliance:report:read'],
-    defaultTitle: defaultReportTitle,
-    modal: true,
-    component: lazy(
-      () =>
-        import(
           /* webpackChunkName: "ReportDetails" */ 'SmartComponents/ExportPDF/ExportPDF'
         ),
     ),
