+++ src/SmartComponents/ExportPDF/ExportPDF.js
@@ -9,13 +9,13 @@ import {
 } from 'PresentationalComponents';
 import useNavigate from '@redhat-cloud-services/frontend-components-utilities/useInsightsNavigate';
 import { useAddNotification } from '@redhat-cloud-services/frontend-components-notifications/hooks';
-import ExportPDFForm from 'SmartComponents/ReportDownload/Components/ExportPDFForm';
-import useExportSettings from 'SmartComponents/ReportDownload/hooks/useExportSettings';
+import ExportPDFForm from './Components/ExportPDFForm';
+import useExportSettings from './hooks/useExportSettings';
 import useReport from 'Utilities/hooks/api/useReport';
 import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 import { exportNotifications } from './constants';
 
-export const ReportDownload = () => {
+export const ExportPDF = () => {
   const addNotification = useAddNotification();
   const { report_id: reportId } = useParams();
   const [loadingPDF, setLoadingPDF] = useState(false);
@@ -113,4 +113,4 @@ export const ReportDownload = () => {
   );
 };
 
-export default ReportDownload;
+export default ExportPDF;
