+++ src/SmartComponents/ExportPDF/ExportPDF.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
-import { ReportDownload } from './ExportPDF';
+import { ExportPDF } from './ExportPDF';
 
 jest.mock('react-router-dom', () => ({
   useParams: jest.fn(() => ({ report_id: 'test-report-id' })),
@@ -11,7 +11,7 @@ jest.mock(
   '@redhat-cloud-services/frontend-components-utilities/useInsightsNavigate',
   () => jest.fn(() => jest.fn()),
 );
-jest.mock('SmartComponents/ReportDownload/hooks/useExportSettings', () =>
+jest.mock('./hooks/useExportSettings', () =>
   jest.fn(() => ({
     exportSettings: {},
     setExportSetting: jest.fn(),
@@ -39,9 +39,9 @@ jest.mock('@redhat-cloud-services/frontend-components/useChrome', () =>
   })),
 );
 
-describe('ReportDownload', () => {
+describe('ExportPDF', () => {
   it('renders the Compliance report modal title', () => {
-    render(<ReportDownload />);
+    render(<ExportPDF />);
 
     expect(screen.getByText('Compliance report'));
     expect(screen.getByText('Test report title'));
