+++ src/SmartComponents/ExportPDF/constants.js
@@ -14,3 +14,12 @@ export const exportNotifications = {
     autoDismiss: false,
   },
 };
+
+export const DEFAULT_EXPORT_SETTINGS = {
+  compliantSystems: false,
+  nonCompliantSystems: true,
+  unsupportedSystems: true,
+  nonReportingSystems: true,
+  topTenFailedRules: true,
+  userNotes: undefined,
+};
