--- src/SmartComponents/ReportDownload/Components/ReportPDF.test.js
@@ -1,10 +1,0 @@
-// import ReportPDF from './ReportPDF';
-
-describe('ReportPDF', () => {
-  // TODO Unskip when mocking of pdf-generator components is possible
-  it.skip('expect to render without error', () => {
-    //     const wrapper = shallow(<ReportPDF />);
-    //
-    //     expect(toJson(wrapper)).toMatchSnapshot();
-  });
-});
