--- src/SmartComponents/ReportDownload/constants.js
@@ -1,8 +1,0 @@
-export const DEFAULT_EXPORT_SETTINGS = {
-  compliantSystems: false,
-  nonCompliantSystems: true,
-  unsupportedSystems: true,
-  nonReportingSystems: true,
-  topTenFailedRules: true,
-  userNotes: undefined,
-};
